+++ core/src/test/java/io/fundrequest/platform/keycloak/KeycloakRepositoryImplTest.java
@@ -1,8 +1,5 @@
-package io.fundrequest.core.keycloak;
+package io.fundrequest.platform.keycloak;
 
-import io.fundrequest.platform.keycloak.KeycloakRepository;
-import io.fundrequest.platform.keycloak.Provider;
-import io.fundrequest.platform.keycloak.UserIdentity;
 import org.junit.Before;
 import org.junit.Test;
 import org.keycloak.admin.client.resource.RealmResource;
@@ -16,7 +13,7 @@ import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class KeycloakRepositoryTest {
+public class KeycloakRepositoryImplTest {
 
     private RealmResource realmResource;
     private KeycloakRepository keycloakRepository;
@@ -24,7 +21,7 @@ public class KeycloakRepositoryTest {
     @Before
     public void setUp() throws Exception {
         realmResource = mock(RealmResource.class, RETURNS_DEEP_STUBS);
-        keycloakRepository = new KeycloakRepository(realmResource, "url");
+        keycloakRepository = new KeycloakRepositoryImpl(realmResource, "url");
     }
 
     @Test
