+++ keycloak/src/main/java/io/fundrequest/platform/keycloak/KeycloakConfig.java
@@ -5,8 +5,10 @@ import org.keycloak.admin.client.resource.RealmResource;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 @Configuration
+@Profile("!local")
 public class KeycloakConfig {
 
     @Value("${keycloak.auth-server-url}")
